+++ app/layout.js
@@ -1,0 +1,18 @@
+import {Inter} from "next/font/google";
+import "./globals.css";
+
+const inter = Inter({subsets: ['latin']});
+
+export const metadata = {
+  title: 'Portfolio',
+  description: 'Generated by create next app'
+}
+
+export default function RootLayout({ children }){
+  return (
+   <html lang="en">
+      <body className={inter.className}>{children}</body>
+   </html>
+  )
+
+}
