+++ package.json
@@ -10,19 +10,18 @@
   },
   "dependencies": {
     "@emailjs/browser": "^3.10.0",
-    "@next/font": "^13.1.6",
     "email-js": "^2.0.2",
     "eslint": "8.33.0",
-    "eslint-config-next": "13.1.6",
     "framer-motion": "^9.0.2",
-    "next": "13.1.6",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "next": "^14.0.4",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.7.1",
     "react-intersection-observer": "^9.4.2"
   },
   "devDependencies": {
     "autoprefixer": "^10.4.13",
+    "eslint-config-next": "^14.0.4",
     "postcss": "^8.4.21",
     "tailwindcss": "^3.2.4"
   }
