--- pages/_document.js
@@ -1,14 +1,0 @@
-import { Html, Head, Main, NextScript } from 'next/document'
-
-export default function Document() {
-  return (
-    <Html lang="en">
-      <Head />
-      <link href="https://fonts.googleapis.com/css2?family=Comfortaa&family=Labrada:wght@500&family=Rubik:wght@600&display=swap" rel="stylesheet"></link>
-      <body>
-        <Main />
-        <NextScript />
-      </body>
-    </Html>
-  )
-}
